+++ api/infrastructure/orm/model/user.go
@@ -6,5 +6,5 @@ type User struct {
 	Name      string `gorm:"size:50;not null"`
 	UID       string `gorm:"not null;unique_index"`
 	APIKey    string `gorm:"size:60;not null;unique_index"`
-	SysRoleID uint   `gorm:"not null"`
+	SysRoleID uint64 `gorm:"not null"`
 }
